requirements.txt
torch
torchvision
torchsummary
pyyaml
numpy
tqdm